Move users models inside accounts app
