--- asso/core/models/users.py
@@ -1,73 +1,0 @@
-from django.contrib.auth.base_user import BaseUserManager
-from django.contrib.auth.models import AbstractUser
-from django.contrib.auth.validators import UnicodeUsernameValidator
-from django.db import models
-from django.utils.translation import gettext_lazy as _
-
-
-class UserManager(BaseUserManager):
-    use_in_migrations = True
-
-    def create_user(self, email, password, **extra_fields):
-        """Create and save a user with the given email and password."""
-        if not email:
-            raise ValueError(_("The Email must be set"))
-        email = self.normalize_email(email)
-        user = self.model(email=email, **extra_fields)
-        user.set_password(password)
-        user.save()
-        return user
-
-    def create_superuser(self, email, password, **extra_fields):
-        """Create and save a SuperUser with the given email and password."""
-        extra_fields.setdefault("is_staff", True)
-        extra_fields.setdefault("is_superuser", True)
-        extra_fields.setdefault("is_active", True)
-
-        if extra_fields.get("is_staff") is not True:
-            raise ValueError(_("Superuser must have is_staff=True."))
-        if extra_fields.get("is_superuser") is not True:
-            raise ValueError(_("Superuser must have is_superuser=True."))
-        return self.create_user(email, password, **extra_fields)
-
-
-class User(AbstractUser):
-    email = models.EmailField(_("email address"), unique=True)
-
-    first_name = models.CharField(
-        _("First Name"),
-        max_length=150,
-        blank=False,
-    )
-    last_name = models.CharField(
-        _("Last Name"),
-        max_length=150,
-        blank=False,
-    )
-
-    username_validator = UnicodeUsernameValidator()
-    username = models.CharField(
-        _("username"),
-        max_length=150,
-        unique=True,
-        null=True,
-        blank=True,
-        help_text=_(
-            "Optional, 150 characters or fewer. Letters, digits and @/./+/-/_ only."
-        ),
-        validators=[username_validator],
-    )
-
-    USERNAME_FIELD = "email"
-    REQUIRED_FIELDS = []
-
-    objects = UserManager()
-
-    @property
-    def full_name(self) -> str:
-        return self.get_full_name()
-
-    def __str__(self):
-        if full_name := self.get_full_name():
-            return f"{full_name} ({self.email})"
-        return self.email
